Pass all malformed requests to Django.
If get_path_info raises a UnicodeDecodeError, it means the URL
was malformed. Django usually handles this with a 400 Bad Request
response. If dj_static is in use, it will instead choke on the
exception, generating a 500 response.

This change makes it so that, if dj_static sees the UnicodeDecodeError,
it will just give up and pass it to Django to handle. We then rely on
Django's behavior (which as-is, will quickly enter the fail path and
return the 400) to handle this malformed request.

Django's behavior seems correct per
https://code.djangoproject.com/ticket/25623?cversion=0&cnum_hist=2